% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_t.R
\name{a_t}
\alias{a_t}
\title{Number of species evolved via anagenesis}
\usage{
a_t(Ana, Nat_tm1, NewArrivals_t, C0, N_t)
}
\arguments{
\item{Ana}{The per-species probability of anagenesis per time step}

\item{Nat_tm1}{Number of natives of the previous time step}

\item{NewArrivals_t}{Number of newly immigrating species}

\item{C0}{The per-species probability of successful colonization at full empty niche space}

\item{N_t}{Empty niches}
}
\value{
Number of species evolved via anagenesis
}
\description{
This function calculates the number of species evolved via anagenesis
(Equation 4 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.

Hauffe, T., D. Delicado, R.S. Etienne and L. Valente (submitted).
Lake expansion increases equilibrium diversity via the target effect of island biogeography
}
\author{
Torsten Hauffe
}
