% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aS_t.R
\name{aS_t}
\alias{aS_t}
\title{Number of new adaptive speciation events}
\usage{
aS_t(Sr_tm1, N_t, S0)
}
\arguments{
\item{Sr_tm1}{Species richness of the island of the previous time step}

\item{N_t}{Empty niches}

\item{S0}{The per species probability of adaptive speciation at maximum empty niche space}
}
\value{
Number of new adaptive speciation events
}
\description{
This function calculates the number of new adaptive speciation events
(Equation 7 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.

Hauffe, T., D. Delicado, R.S. Etienne and L. Valente (submitted).
Lake expansion increases equilibrium diversity via the target effect of island biogeography.
}
\author{
Torsten Hauffe
}
