% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_t.R
\name{end_t}
\alias{end_t}
\title{Endemic richness}
\usage{
end_t(ExtinctProb_t, End_tm1, A_t, Em_t, NaS_t, AS_t)
}
\arguments{
\item{ExtinctProb_t}{Probability for a species to go extinct in a time step}

\item{End_tm1}{Endemic richness of the previous time step}

\item{A_t}{Number of species evolved via anagenesis}

\item{NaS_t}{Number of new non-adaptive speciation events}

\item{AS_t}{Number of new adaptive speciation events}
}
\value{
Endemic richness
}
\description{
This function calculates the endemic richness
(Equation 12 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
