% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extinct_Prob_t.R
\name{extinctProb_t}
\alias{extinctProb_t}
\title{Probability for a species to go extinct in a time step}
\usage{
extinctProb_t(N_t, E0)
}
\arguments{
\item{N_t}{Empty niches}

\item{E0}{The per-species probability of extinction when richness equals Kmax}
}
\value{
Probability for a species to go extinct in a time step
}
\description{
This function calculates the probability for a species to go extinct in a time step
(Equation 8 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.

Hauffe, T., D. Delicado, R.S. Etienne and L. Valente (submitted).
Lake expansion increases equilibrium diversity via the target effect of island biogeography
}
\author{
Torsten Hauffe
}
