% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmax_t.R
\name{kmax_t}
\alias{kmax_t}
\title{Carrying capacity of the island}
\usage{
kmax_t(Ms, Elevation_t, Topography_t, Area_t, Z, Ma)
}
\arguments{
\item{Ms}{Species richness    of the mainland    source pool}

\item{Elevation_t}{Elevation}

\item{Topography_t}{Topography}

\item{Area_t}{Area}

\item{Z}{The exponent of the species-area relationship between the island and the source area}

\item{Ma}{The realized size of the source area that provides new species}
}
\value{
Carrying capacity of the island
}
\description{
This function calculates the carrying capacity of the island
(Equation 9 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\seealso{
\code{\link{calc_diversity}}
}
\author{
Torsten Hauffe
}
