% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_t.R
\name{n_t}
\alias{n_t}
\title{Number of empty niches}
\usage{
n_t(Kmax_t, Sr_t, Ms)
}
\arguments{
\item{Kmax_t}{Carrying capacity}

\item{Sr_t}{Total island species richness}

\item{Ms}{Species richness    of the mainland    source pool}
}
\value{
Number of empty niches
}
\description{
This function calculates the number of empty niches
(Equation 10 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
