% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naS_t.R
\name{naS_t}
\alias{naS_t}
\title{Number of new non-adaptive speciation events}
\usage{
naS_t(Sr_tm1, Topography_t, Vs)
}
\arguments{
\item{Sr_tm1}{Species richness of the island of the previous time step}

\item{Topography_t}{Topography}

\item{Vs}{Probability of a vicariance event when topography equals 1}
}
\value{
Number of new non-adaptive speciation events
}
\description{
This function calculates the number of new non-adaptive speciation events
(Equation 6 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
