% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nat_t.R
\name{nat_t}
\alias{nat_t}
\title{Native richness}
\usage{
nat_t(ExtinctProb_t, Nat_tm1, NewColonizers_t, A_t, Em_t)
}
\arguments{
\item{ExtinctProb_t}{Probability for a species to go extinct in a time step}

\item{Nat_tm1}{Species richness of the island of the previous time step}

\item{NewColonizers_t}{Number of new successfully colonizing species}

\item{A_t}{Number of species evolved via anagenesis}

\item{Em_t}{Number of emigration events}
}
\value{
Native richness
}
\description{
This function calculates the native richness
(Equation 11 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
