% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newArrivals_t.R
\name{newArrivals_t}
\alias{newArrivals_t}
\title{Number of newly immigrating species}
\usage{
newArrivals_t(Imm, Ms, Iso_t, NewSpecies_t, Area_t = 1,
  Elevation_t = 1)
}
\arguments{
\item{Imm}{The probability of one species arriving from the mainland per time step}

\item{Ms}{Species richness    of the mainland    source pool}

\item{Iso_t}{A descriptor of island isolation in arbitrary units}

\item{NewSpecies_t}{Probability that an immigrant species is not already on the island}

\item{Area_t}{Island area at moment t, to simulate the target effect (not included in Borregaard 2016)}

\item{Elevation_t}{Island elevation at moment t, to simulate the target effect (not included in Borregaard 2016)}
}
\value{
Number of newly immigrating species
}
\description{
This function calculates the number of newly immigrating species
(Equation 2 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M.K., T.J. Matthews and R.J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.

Hauffe, T., D. Delicado, R.S. Etienne and L. Valente (submitted).
Lake expansion increases equilibrium diversity via the target effect of island biogeography
}
\author{
Torsten Hauffe
}
