% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newColonizers_t.R
\name{newColonizers_t}
\alias{newColonizers_t}
\title{Number of new successfully colonizing species}
\usage{
newColonizers_t(NewArrivals_t, C0, N_t)
}
\arguments{
\item{NewArrivals_t}{Number of newly immigrating species}

\item{C0}{The per-species probability of successful colonization at full empty niche space}

\item{N_t}{Empty niches}
}
\value{
Number of new successfully colonizing species
}
\description{
This function calculates the number of new successfully colonizing species
(Equation 3 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
