% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newSpecies_t.R
\name{newSpecies_t}
\alias{newSpecies_t}
\title{Probability that an immigrant species is not already on the island}
\usage{
newSpecies_t(Nat_tm1, Ms, Msa)
}
\arguments{
\item{Nat_tm1}{Island natives of the previous time step}

\item{Ms}{Species richness    of the mainland    source pool}

\item{Msa}{Power exponent controlling the abundance distribution of the source}
}
\value{
Probability that an immigrant species is not already on the island
}
\description{
This function calculates the probability	that an immigrant species is not already on the island
(Equation 1 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M.K., T.J. Matthews and R.J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
