% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_t.R
\name{sr_t}
\alias{sr_t}
\title{Total island species richness}
\usage{
sr_t(Nat_t, End_t)
}
\arguments{
\item{Nat_t}{Native richness}

\item{End_t}{Endemic richness}
}
\value{
Total island species richness
}
\description{
This function calculates the total island species richness
(Equation 12 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
