% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_t.R
\name{em_t}
\alias{em_t}
\title{Number of emigration events}
\usage{
em_t(End_tm1, Emi, Iso_t)
}
\arguments{
\item{Ana}{The per-species probability of anagenesis per time step}

\item{Nat_tm1}{Number of natives of the previous time step}

\item{NewArrivals_t}{Number of newly immigrating species}
}
\value{
Number of emigration events
}
\description{
This function calculates the number of emigration events
(Equation 5 in Appendix 2 of Borregaard et al. 2016)
}
\references{
Borregaard, M. K., T. J. Matthews and R. J. Whittaker (2016).
The general dynamic model: towards a unified theory of island biogeography?
Global Ecology and Biogeography, 25(7), 805-816.
}
\author{
Torsten Hauffe
}
